% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_df.R
\name{ctr_df}
\alias{ctr_df}
\title{Creates a data.frame with the degree of each layer's nodes}
\usage{
ctr_df (g.list, layer_names)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{layer_names}{Vector with the name of each layer.}
}
\value{
Returns a data frame with degree's value, phylum and specie of each node.
}
\description{
Creates a data.frame with the degree of each layer's nodes
}
