% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_g.R
\name{ctr_g}
\alias{ctr_g}
\title{This function assigns colors to the network nodes based on their centrality measurements.}
\usage{
ctr_g (g, ctr_type)
}
\arguments{
\item{g}{An igraph object.}

\item{ctr_type}{Centrality type ("degree", "betweenness" or "closeness").}
}
\value{
An igraph object with colored nodes.
}
\description{
This function assigns colors to the network nodes based on their centrality measurements.
}
\examples{

data (net_dstoolA)
ctr_g (net_dstoolA, ctr_type = "degree")
ctr_g (net_dstoolA, ctr_type = "closeness")

data (net_dstoolB)
ctr_g (net_dstoolB, ctr_type = "betweenness")
ctr_g (net_dstoolB, ctr_type = "degree")
}
