% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_ml.R
\name{ctr_ml}
\alias{ctr_ml}
\title{Colors the nodes according to their centrality.}
\usage{
ctr_ml (g.list, ctr_type)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{ctr_type}{Centrality type ("degree", "betweenness", "closeness").}
}
\value{
A list of igraph objects which
}
\description{
Colors the nodes according to their centrality.
}
\examples{
# 1+1

}
