% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_plot.R
\name{diff_plot}
\alias{diff_plot}
\title{This function identifies the most abundant taxa and distinguishes between nodes with statistically significant differences in their abundance (red) and those without (blue). The size of nodes corresponds to abundances.}
\usage{
diff_plot(diff_g, T_collapsed)
}
\arguments{
\item{diff_g}{igraph object obtained with diff_nodes_graph function.}

\item{T_collapsed}{Abundance table collapsed at a certain taxonomic level (preprocessed by T_collapse).}
}
\value{
Returns a plot with both networks collapsed.
}
\description{
This function identifies the most abundant taxa and distinguishes between nodes with statistically significant differences in their abundance (red) and those without (blue). The size of nodes corresponds to abundances.
}
\examples{

data("BN_nets")
data("beetle_nightshade")
BN_genus <- T_collapse(is_phyloseq = T, ps = beetle_nightshade, names_level = "Genus")
# Insect and plant data isolation
library("phyloseq")
Insect<-which(sample_data(beetle_nightshade)$Type =="Insect")
Insect<-sample_data(beetle_nightshade)$ID[Insect]; Insect
Plant<-which(sample_data(beetle_nightshade)$Type =="Plant")
Plant<-sample_data(beetle_nightshade)$ID[Plant]; Plant
Insectmat <- BN_genus[Insect,]
Plantmat <- BN_genus[Plant,]
# Network inference
insect_sparcc<-net_inference(Insectmat, "sparcc", p = 0.4)
plant_sparcc<-net_inference(Plantmat, "sparcc", p = 0.4)
# Significant differences between replica nodes
mat_list<-list(Insectmat, Plantmat)
g.list<-list(insect_sparcc, plant_sparcc)
diff_g<-diff_nodes_graph(BN_genus, n = 20, mat_list, g.list, alpha = 0.05)
diff_plot(diff_g = diff_g, T_collapsed = BN_genus)

}
