% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_properties.R
\name{ml_properties}
\alias{ml_properties}
\title{Quantifies different topological properties of a multilayer network.}
\usage{
ml_properties (g.list, treatments)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{treatments}{A vector with the names of each layer.}
}
\value{
The function returns a data.frame, which contains topological metrics of the networks.
}
\description{
Quantifies different topological properties of a multilayer network.
}
\examples{

mlDstool <- readRDS (file = "data/Dstool_mlnet.rds")
ntreat <- c("before_travel", "traveling")
ml_properties (mlDstool, ntreat)
}
