% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_TaxGroup.R
\name{ml_TaxGroup}
\alias{ml_TaxGroup}
\title{Assigns the taxonomic membership of the nodes.}
\usage{
ml_TaxGroup(g.list, T_table, g_tax, p_tax)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{T_table}{Taxa table.}

\item{g_tax}{description}

\item{p_tax}{description}
}
\value{
n
}
\description{
Assigns the taxonomic membership of the nodes.
}
\examples{
n
}
