% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_inference.R
\name{net_inference}
\alias{net_inference}
\title{This function infers a co-abundance network from an abundance table.}
\usage{
net_inference (taxa_abs, method)
}
\arguments{
\item{taxa_abs}{An abundance table (data.frame) preprocessed with function T_collapse.}

\item{method}{The method to create the network ("sparcc" or "aracne").}

\item{p}{When method="sparcc", a threshold between 0-1.}
}
\value{
The function returns an igraph object representing the existing co-abundances between nodes at a certain taxonomic level.
}
\description{
This function infers a co-abundance network from an abundance table.
}
\examples{

data("beetle_nightshade")
BN_genus <- T_collapse(is_phyloseq = T, ps = beetle_nightshade, names_level = "Genus")
# Insect data isolation
library(phyloseq)
Insect<-which(sample_data(beetle_nightshade)$Type =="Insect")
Insect<-sample_data(beetle_nightshade)$ID[Insect]
Insectmat <- BN_genus[Insect,]
# Network inference
net_inference(Insectmat, "sparcc", p = 0.4)

}
