% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_plot.R
\name{net_plot}
\alias{net_plot}
\title{Creates a graph of the network.}
\usage{
net_plot (g, type)
}
\arguments{
\item{g}{An igraph object preprocessed by v_colored or ctr_g.}

\item{type}{You can choose between "phylo" or "centrality".}
}
\value{
An igraph object.
}
\description{
Creates a graph of the network.
}
\examples{

g <- readRDS("examples/Dstool_net.rds")
net_plot (g, type = "centrality")
net_plot (g, type = "phylo")

j <- readRDS (file="examples/net_dietswap.rds")
net_plot (j, type = "centrality")
net_plot (j, type = "phylo")
}
