% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_color_mat.R
\name{node_color_mat}
\alias{node_color_mat}
\title{Creates a matrix that contains assigned colors for the nodes of a multilayer network.}
\usage{
node_color_mat (g.list, type)
}
\arguments{
\item{g.list}{List of igraph objects.}

\item{type}{"phylo" assigns the colors according to a certain taxonomic level; "centrality" assings the colors according to the centrality level.}
}
\value{
Returns a heatmap of the multilayer network according the type you chose.
}
\description{
Creates a matrix that contains assigned colors for the nodes of a multilayer network.
}
\examples{

data ("mlnet_dstoolAB")

node_color_mat (mlnet_dstoolAB, type = "phylo")

node_color_mat (mlnet_dstoolAB, type = "centrality")
}
