% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_R2.R
\name{slope_R2}
\alias{slope_R2}
\title{This function computes  the slope and R-squared values of a linear-regression for topological properties from networks.}
\usage{
slope_R2 (df_properties)
}
\arguments{
\item{df_properties}{A data.frame generated by ml_properties.}
}
\value{
Returns a data.frame describing the slope and the R-squared of linear models.
}
\description{
This function computes  the slope and R-squared values of a linear-regression for topological properties from networks.
}
\examples{

data ("ml_prop_dstool")

slope_R2 (ml_prop_dstool)
}
