% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_collapse.R
\name{T_collapse}
\alias{T_collapse}
\title{Adds the abundances at a certain taxonomic level}
\usage{
T_collapse (T_table, O_table, names_level)
}
\arguments{
\item{ps}{A phyloseq object}

\item{T_table}{A taxa table}

\item{O_table}{An OTU table (ASV?)}

\item{names_level}{Taxonomic level to separate _}
}
\value{
Returns a class array matrix at a certain taxonomic level.
}
\description{
Adds the abundances at a certain taxonomic level
}
\examples{
1+1
}
