% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_mat.R
\name{abs_mat}
\alias{abs_mat}
\title{Colors nodes according to their abundance.}
\usage{
abs_mat(abs.list, g.list, n)
}
\arguments{
\item{abs.list}{Abundance list}

\item{g.list}{List of igraph objects (multilayer network).}

\item{n}{A vertex}
}
\value{
Returns an igraph object.
}
\description{
Colors nodes according to their abundance.
}
\examples{

dstool_net <- readRDS(file = "data/Dstool_mlnet.rds")
abs_mat(david_stool_lineage, dstool_net, 2)
}
