% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_mat.R
\name{abs_mat}
\alias{abs_mat}
\title{Colors nodes according their abundance.}
\usage{
abs_mat(abs.list, g.list, n)
}
\arguments{
\item{abs.list}{Abundance list}

\item{g.list}{List of igraph objects (multilayer network).}

\item{n}{A vertex}
}
\value{
Information
}
\description{
Colors nodes according their abundance.
}
\examples{
# 1+1
}
