% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_mat.R
\name{abs_mat}
\alias{abs_mat}
\title{This function colors nodes by abundance at a taxonomic level.}
\usage{
abs_mat (abs.list, g.list, n)
}
\arguments{
\item{abs.list}{Abundance list, the same lists you used to generate the network.}

\item{g.list}{List of igraph objects (multilayer network).}

\item{n}{Vertex number. Notice that if n>20 the nodes won't show too big.}
}
\value{
Returns a matrix, each column is a layer, and the values of the rows corresponds to the size of each node.
}
\description{
This function colors nodes by abundance at a taxonomic level.
}
\examples{

data (mlnet_dstoolAB)

abs_mat (abs.list = david_stool_lineages, g.list = mlnet_dstoolAB, n=10)
}
