% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_degree_abs.R
\name{cor_degree_abs}
\alias{cor_degree_abs}
\title{Graphs the correlation between degree and abundances.}
\usage{
cor_degree_abs(g.list, abs.list, layer_names, layer_colors, title)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{abs.list}{List of abundances tables.}

\item{layer_names}{The names of each layer of the network (ex. "control layer").}

\item{layer_colors}{Assign a color to represent the layer (ex. "pink")}

\item{title}{Plot's title.}
}
\value{
Returns a correlation plot between the degree and abundance of each node. It shows the R-square and p-value.
}
\description{
Graphs the correlation between degree and abundances.
}
\examples{

# Load multilayer net
g <- readRDS (file = "examples/Dstool_mlnet.rds")

cor_degree_abs (g, david_stool_lineages, layer_names = c("before", "meanwhile", "after"), layer_colors = c("blue", "purple", "green"), title = "Example")

}
