% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_degree_abs.R
\name{cor_degree_abs}
\alias{cor_degree_abs}
\title{Correlation between degree and abundance
Graphing Pearson correlation between degree and abundance.}
\usage{
cor_degree_abs(g.list, abs.list, layer_names, layer_colors, title)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{abs.list}{List of abundances tables.}

\item{layer_names}{The names of each layer of the network (ex. "control layer").}

\item{layer_colors}{Assign a color to represent the layer (ex. "pink")}

\item{title}{Plot's title.}
}
\value{
Returns a correlation plot between the degree and abundance of each node. It shows the R-square and p-value.
}
\description{
Correlation between degree and abundance
Graphing Pearson correlation between degree and abundance.
}
\examples{

data("BN_nets")
data("beetle_nightshade")
BN_genus <- T_collapse(is_phyloseq = T, ps = beetle_nightshade, names_level = "Genus")
# Insect and plant data isolation
library("phyloseq")
Insect<-which(sample_data(beetle_nightshade)$Type =="Insect")
Insect<-sample_data(beetle_nightshade)$ID[Insect]; Insect
Plant<-which(sample_data(beetle_nightshade)$Type =="Plant")
Plant<-sample_data(beetle_nightshade)$ID[Plant]; Plant
Insectmat <- BN_genus[Insect,]
Plantmat <- BN_genus[Plant,]
# Network inference
insect_sparcc<-net_inference(Insectmat, "sparcc", p = 0.4)
plant_sparcc<-net_inference(Plantmat, "sparcc", p = 0.4)
# Degree vs. abundance plot
BN_ml<-list(insect_sparcc, plant_sparcc)
BN_abs<-list(Insectmat, Plantmat)
cor_degree_abs(g.list = BN_ml, abs.list = BN_abs, layer_names = c("Beetle", "Nightshade"), layer_colors = c("red4", "green4"), title = "Beetle-nightshade")

}
