% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_degree_abs.R
\name{cor_degree_abs}
\alias{cor_degree_abs}
\title{Graph the correlation between degree and abundances.}
\usage{
cor_degree_abs(g.list, abs.list, layer_names, layer_colors, title)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{abs.list}{List of abundances tables.}

\item{layer_names}{description}

\item{layer_colors}{description}

\item{title}{description}
}
\value{
n
}
\description{
Graph the correlation between degree and abundances.
}
\examples{
1+1

}
