% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_df.R
\name{ctr_df}
\alias{ctr_df}
\title{The function ctr_df generates a data.frame containing the degree of nodes for each layer.}
\usage{
ctr_df (g.list, layer_names)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network), preprocessed by ml_TaxGroup.}

\item{layer_names}{Vector with the name of each layer.}
}
\value{
Returns a data.frame with centrality value (if preprocessed with ctr_ml or ctr_g: degree, betweenness or closeness), p_tax and g_tax of each node (the function takes this info from ml_TaxGroup).
}
\description{
The function ctr_df generates a data.frame containing the degree of nodes for each layer.
}
\examples{

#Load the example net
data ("mlTaxGroup_dstool")

# Create a vector with layers' names
v <- c("sample A", "sample B")

ctr_df (mlTaxGroup_dstool, layer_names = v)
}
