% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_df.R
\name{ctr_df}
\alias{ctr_df}
\title{Creates a data.frame with the degree of each layer's nodes.}
\usage{
ctr_df (g.list, layer_names)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network), preprocessed by ml_TaxGroup.}

\item{layer_names}{Vector with the name of each layer.}
}
\value{
Returns a data.frame with centrality value (if preprocessed with ctr_ml or ctr_g: degree, betweenness or closeness), p_tax and g_tax of each node (function take this info from TaxGroup_ml).
}
\description{
Creates a data.frame with the degree of each layer's nodes.
}
\examples{

#Load the example net
g <- readRDS (file = "examples/Dstool_mlnet.rds")

# Assign layer names
v <- c("before", "meanwhile", "after")

ctr_df (g, layer_names = v)
}
