% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_g.R
\name{ctr_g}
\alias{ctr_g}
\title{Colors the nodes of a network according to their centrality.}
\usage{
ctr_g(g, ctr_type)
}
\arguments{
\item{g}{An igraph object.}

\item{ctr_type}{Centrality type ("degree", "betweenness" or "closeness").}
}
\value{
An igraph object with
}
\description{
Colors the nodes of a network according to their centrality.
}
\examples{
# 1+1

}
