% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_nodes_graph.R
\name{diff_nodes_graph}
\alias{diff_nodes_graph}
\title{This function identifies the most abundant taxa and distinguishes between nodes with statistically significant differences in their abundance (red) and those without (blue).}
\usage{
diff_nodes_graph(T_Collapsed, n, mat_list, g.list, alpha)
}
\arguments{
\item{n}{Number of nodes to consider.}

\item{mat_list}{List of abundance matrices (preprocessed by T_collapse).}

\item{g.list}{List of igraph objects (multilayer network), just 2 layers.}

\item{alpha}{Significance. It must be a number between 0 and 1.}

\item{T_collapsed}{Abundance table collapsed at a certain taxonomic level (preprocessed by T_collapse).}
}
\value{
Returns a plot with both networks collapsed.
}
\description{
This function identifies the most abundant taxa and distinguishes between nodes with statistically significant differences in their abundance (red) and those without (blue).
}
\examples{

data (tc_dstool)
data (mlnet_dstool)

diff_nodes_graph (tc_dstool, n=10, mat_list=t, g.list=mlnet_dstool, alpha=0.5)

}
