% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_nodes_graph.R
\name{diff_nodes_graph}
\alias{diff_nodes_graph}
\title{Description}
\usage{
diff_nodes_graph(T_Collapsed, n, mat_list, g.list, alpha)
}
\arguments{
\item{n}{description}

\item{mat_list}{description}

\item{g.list}{List of igraph objects (multilayer network).}

\item{alpha}{description}

\item{T_collapsed}{description}
}
\value{
n
}
\description{
Description
}
\examples{
1+1

}
