% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_abundance.R
\name{g_abundance}
\alias{g_abundance}
\title{This function assigns colors to the nodes based on their relative abundance.}
\usage{
g_abundance (layer_mat, g)
}
\arguments{
\item{layer_mat}{Abundance data.frame.}

\item{g}{An igraph object.}
}
\value{
A graph with colored nodes.
}
\description{
This function assigns colors to the nodes based on their relative abundance.
}
\examples{

data ()

g_abundance (david_stool_lineages, net_dstA)

}
