% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_abundance.R
\name{g_abundance}
\alias{g_abundance}
\title{Description}
\usage{
g_abundance(layer_mat, g)
}
\arguments{
\item{layer_mat}{description}

\item{g}{An igraph obejct.}
}
\value{
k
}
\description{
Description
}
\examples{
# k

}
