% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_abundance.R
\name{g_abundance}
\alias{g_abundance}
\title{Assign a color to each node depending of the relative abundance.}
\usage{
g_abundance(layer_mat, g)
}
\arguments{
\item{layer_mat}{Abundance data.frame.}

\item{g}{An igraph obejct.}
}
\value{
k
}
\description{
Assign a color to each node depending of the relative abundance.
}
\examples{
# k

}
