% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2FC.R
\name{log2fc}
\alias{log2fc}
\title{Quantifies and compares Log Fold Change value between chosen layers.}
\usage{
log2fc (phyl_ctr_df, control_layer, test_layer)
}
\arguments{
\item{phyl_ctr_df}{A data.frame preprocessed with phyl_ctr_df function.}

\item{control_layer}{"Name" of control layer.}

\item{test_layer}{"Name" of test layer.}
}
\value{
Returns a Log Fold Change plot of the centrality between layers.
}
\description{
Quantifies and compares Log Fold Change value between chosen layers.
}
\examples{

library (viridis)
data("phyl_ctr_dstool")

log2fc (phyl_ctr_dstool, control_layer = "sample A", test_layer = "sample B")
}
