% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2FC.R
\name{log2fc}
\alias{log2fc}
\title{Description}
\usage{
log2fc(phyl_ctr_df, control_layer, test_layer)
}
\arguments{
\item{phyl_ctr_df}{A data.frame preprocessed with phyl_ctr_df function.}

\item{control_layer}{"Name" of control layer.}

\item{test_layer}{"Name" of test layer.}
}
\value{
Returns a plot of Log Fold Change of the centrality between layers.
}
\description{
Description
}
\examples{
phcDstool <- readRDS(file = "data/Dstool_phyl_ctr.rds")
log2fc (phcDstool, "after_travel", "traveling")
}
