% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2FC.R
\name{log2fc}
\alias{log2fc}
\title{Description}
\usage{
log2fc(phyl_ctr_df, control_layer, test_layer)
}
\arguments{
\item{phyl_ctr_df}{n}

\item{control_layer}{description}

\item{test_layer}{description}
}
\value{
n
}
\description{
Description
}
\examples{
1+1

}
