% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_properties.R
\name{ml_properties}
\alias{ml_properties}
\title{Description.}
\usage{
ml_properties(g.list, treatments)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{treatments}{A vector which contain the name of each layer.}
}
\value{
The function returns a linear regression of each layer.
}
\description{
Description.
}
\examples{

mlDstool <- readRDS (file = "data/Dstool_mlnet.rds")
ntreat <- c("Control", "post-sickness")
ml_properties (mlDstool, ntreat)

}
