% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_properties.R
\name{ml_properties}
\alias{ml_properties}
\title{Multilayer network properties}
\usage{
ml_properties (g.list, treatments)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{treatments}{A vector with the names of each layer.}
}
\value{
The function returns a data.frame, which contains networks' topological metrics.
}
\description{
Quantifies different topological properties of a multilayer network.
}
\examples{

data ("mlnet_dstoolAB")

# The vector with layers' names
t <- c("before", "after")

ml_properties (mlnet_dstoolAB, treatments = t)
}
