% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_properties.R
\name{ml_properties}
\alias{ml_properties}
\title{Description.}
\usage{
ml_properties(g.list, treatments)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{treatments}{description}
}
\value{
Linear regression of each layer.
}
\description{
Description.
}
\examples{
n

}
