% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_TaxGroup.R
\name{ml_TaxGroup}
\alias{ml_TaxGroup}
\title{Assigns the taxonomic membership of the nodes.}
\usage{
ml_TaxGroup (g.list, T_table, g_tax, p_tax)
}
\arguments{
\item{g.list}{List of igraph objects (multilayer network).}

\item{T_table}{Taxa table.}

\item{g_tax}{The taxonomic level at which the nodes will be assigned.}

\item{p_tax}{The taxonomic level at which the network was built.}
}
\value{
The function returns a g.list (multilayer network) with each node named according its taxonomic level.
}
\description{
Assigns the taxonomic membership of the nodes.
}
\examples{

# Load g.list
mlDstool <- readRDS (file="examples/Dstool_mlnet.rds")
# Replace columns' names
colnames (david_stool_lineages) <- c("ID", "Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species", "Summary", "ID")

ml_TaxGroup (mlDstool, T_table=david_stool_lineages, g_tax="Phylum", p_tax="Family")
}
