\name{net_inference}
\alias{ctr_ml}
\title{ctr_ml}
\usage{
ctr_ml(g.list, ctr_type)
}
\description{
Colors the nodes according to their centrality.
}
\arguments{
\item{g.list}{List of \emph{igraph} objects.}

\item{ctr_type}{Centrality type ("degree", "betweenness" or "closeness").}
}

\examples{
}
