% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_inference.R
\name{net_inference}
\alias{net_inference}
\title{Colors the nodes according to their centrality.}
\usage{
net_inference (taxa_abs, method, p)
}
\arguments{
\item{taxa_abs}{An abundance table, preprocessed with function T_collapse.}

\item{method}{The method to create the network ("sparcc", "aracne")}

\item{p}{When method="sparcc", a threshold between 0-1.}
}
\value{
Description
}
\description{
Colors the nodes according to their centrality.
}
\examples{

sparcc_net <- net_inference(a, "sparcc", 0.4)
multilayer_net <- load ("data/obj_ml.RData")
plot (multilayer_net)

aracne_net <- net_inference(a, "aracne")
aracne_obj <- load ("data/obj_aracne.RData")
plot(aracne_obj)
}
