% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_inference.R
\name{net_inference}
\alias{net_inference}
\title{Colors the nodes according to their centrality.}
\usage{
net_inference (taxa_abs, method)
}
\arguments{
\item{taxa_abs}{An abundance table, preprocessed with function T_collapse.}

\item{method}{The method to create the network ("sparcc", "aracne")}

\item{p}{When method="sparcc", a threshold between 0-1.}
}
\value{
The function returns an igraph object..
}
\description{
Colors the nodes according to their centrality.
}
\examples{

object_tc <- readRDS("data/object_tc_dietswap.rds")
sparccnet <- net_inference (object_tc, "sparcc", 0.4)
plot (sparccnet)


object_tc2 <- readRDS("data/object_tc_dstool.rds")
aracnet <- net_inference (object_tc2, "aracne")
plot (aracnet)

}
