% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_plot.R
\name{net_plot}
\alias{net_plot}
\title{Creates a graph of the network.}
\usage{
net_plot (g, type)
}
\arguments{
\item{g}{An igraph object preprocessed by v_colored or ctr_g.}

\item{type}{You can choose between "phylo" or "centrality".}
}
\value{
An igraph object.
}
\description{
Creates a graph of the network.
}
\examples{

data ("ctr_g_dstool")
net_plot (ctr_g_dstool, type = "centrality")

}
