% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_plot.R
\name{net_plot}
\alias{net_plot}
\title{This function creates a network plot. This function also provides options for customizing network visualizations, allowing users to adjust node sizes, edge weights, and color schemes to highlight specific patterns or relationships of interest.}
\usage{
net_plot (g, type)
}
\arguments{
\item{g}{An igraph object preprocessed by v_colored or ctr_g.}

\item{type}{You can choose between "phylo" or "centrality".}
}
\value{
An igraph object.
}
\description{
This function creates a network plot. This function also provides options for customizing network visualizations, allowing users to adjust node sizes, edge weights, and color schemes to highlight specific patterns or relationships of interest.
}
\examples{

data ("ctr_g_dstool")
net_plot (ctr_g_dstool, type = "centrality")

}
