% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_plot.R
\name{net_plot}
\alias{net_plot}
\title{Create a graph of the network.}
\usage{
net_plot(g, type)
}
\arguments{
\item{g}{An igraph object preprocessed by v_colored or ctr_g functions.}

\item{type}{You can choose between "phylo" or "centrality".}
}
\value{
An igraph object..
}
\description{
Create a graph of the network.
}
\examples{

graph <- readRDS("data/dstool_net.rds")
net_plot (graph, type = "centrality")
net_plot (graph, type = "phylo")
}
