% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyl_ctr_df.R
\name{phyl_ctr_df}
\alias{phyl_ctr_df}
\title{Creates a data.frame of the sums of centrality values at a chosen taxonomic level.}
\usage{
phyl_ctr_df (ctr_df, layer_names, n_layers)
}
\arguments{
\item{ctr_df}{Data.frame pre-processed with ctr_df function.}

\item{layer_names}{Name of the layers.}

\item{n_layers}{Number of layers.}
}
\value{
A data.frame.
}
\description{
Creates a data.frame of the sums of centrality values at a chosen taxonomic level.
}
\examples{

Dstool_ctr <- readRDS (file="examples/Dstool_ctr_df.rds")

n <- c("before_sickness", "after_sickness")
phyl_ctr_df (Dstool_ctr, layer_names=n, n_layers=2)
}
