% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyl_ctr_df.R
\name{phyl_ctr_df}
\alias{phyl_ctr_df}
\title{Creates a data.frame of the sums of centrality values at a chosen taxonomic level}
\usage{
phyl_ctr_df(ctr_df, layer_names, n_layers)
}
\arguments{
\item{ctr_df}{Data.frame}

\item{layer_names}{Name of the layers}

\item{n_layers}{Number of layers.}
}
\description{
Creates a data.frame of the sums of centrality values at a chosen taxonomic level
}
