% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyl_ctr_df.R
\name{phyl_ctr_df}
\alias{phyl_ctr_df}
\title{Creates a data.frame of the sums of centrality values at a chosen taxonomic level.}
\usage{
phyl_ctr_df (ctr_df, layer_names, n_layers)
}
\arguments{
\item{ctr_df}{Data.frame pre-processed with ctr_df function.}

\item{layer_names}{Name of the layers.}

\item{n_layers}{Number of layers.}
}
\value{
A data.frame with one column per layer with their centrality value.
}
\description{
Creates a data.frame of the sums of centrality values at a chosen taxonomic level.
}
\examples{

data (ctr_df_dstool)

# Create a vector with layers' names
n <- c("sample A", "sample B")

phyl_ctr_df (ctr_df_dstool, layer_names = n, n_layers = 2)
}
