% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyl_ctr_plot.R
\name{phyl_ctr_plot}
\alias{phyl_ctr_plot}
\title{Description}
\usage{
phyl_ctr_plot(phyl_ctr_df, layer_name, title)
}
\arguments{
\item{phyl_ctr_df}{A data.frame preprocessed by phyl_ctr_df}

\item{layer_name}{Name of the layer to graph/plot.}

\item{title}{Name for the graph.}
}
\value{
description
}
\description{
Description
}
\examples{

phyl_ds <- readRDS(file = "phyl_Dstool.rds")
phyl_ctr_plot (phyl_ds, "Control", "Example")
}
