% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyl_ctr_plot.R
\name{phyl_ctr_plot}
\alias{phyl_ctr_plot}
\title{This function shows centrality importance of the phyla.}
\usage{
phyl_ctr_plot(phyl_ctr_df, layer_name, title)
}
\arguments{
\item{phyl_ctr_df}{A data.frame preprocessed by phyl_ctr_df}

\item{layer_name}{Name of the layer to graph/plot.}

\item{title}{Name for the graph.}
}
\value{
A bar plot of degree importance for each phylum.
}
\description{
This function shows centrality importance of the phyla.
}
\examples{

phyl_ds <- readRDS(file = "phyl_Dstool.rds")
phyl_ctr_plot (phyl_ds, "before_travel", "after_travel")
}
