% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_collapse.R
\name{T_collapse}
\alias{T_collapse}
\title{Adds the abundances at a certain taxonomic level.}
\usage{
T_collapse (is_phyloseq=T, ps=..., names_level=...)
}
\arguments{
\item{ps}{If "is_phyloseq = T" A phyloseq object.}

\item{T_table}{(If is_phyloseq = F) You have to introduce a taxa table.}

\item{O_table}{(If is_phyloseq = F) An OTUs table.}

\item{names_level}{(If is_phyloseq = F) Taxonomic level to separate.}
}
\value{
Returns a class array matrix at a certain taxonomic level.
}
\description{
Adds the abundances at a certain taxonomic level.
}
\examples{
# With BiocManager, package "microbiome"
# if is_phyloseq = T

library (microbiome)
data ("dietswap")
T_collapse (is_phyloseq = T, ps = dietswap, names_level ="Phylum")

# if is_phyloseq = F

library(seqtime)
colnames(david_stool_lineages)<- c("ID", "Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species", "Summary", "ID")

T_collapse(F, T_table = david_stool_lineages, O_table = david_stoolA_otus, names_level = "Genus")
}
