% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxGroup.R
\name{TaxGroup}
\alias{TaxGroup}
\title{Description}
\usage{
TaxGroup(g, T_table, g_tax, p_tax)
}
\arguments{
\item{g}{An igraph object.}

\item{T_table}{Taxa table.}

\item{g_tax}{description}

\item{p_tax}{description}
}
\value{
description
}
\description{
Description
}
\examples{
1+1

}
