% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxGroup.R
\name{TaxGroup}
\alias{TaxGroup}
\title{Identifies the membership of a node to a certain higher-ranking taxon.}
\usage{
TaxGroup(g, T_table, g_tax, p_tax)
}
\arguments{
\item{g}{An igraph object.}

\item{T_table}{Taxa table as data.frame}

\item{g_tax}{The taxonomic level at which the nodes will be colored.}

\item{p_tax}{The taxonomic level at which the network was built.}
}
\value{
Returns an igraph object with nodes colored according to their taxonomic rank.
}
\description{
Identifies the membership of a node to a certain higher-ranking taxon.
}
\examples{

# The input can be a net pre-processed by net_inference
# The data is david_stool

tcollapsed <- readRDS (file="examples/object_tc_dstool.rds")
g <- net_inference (tcollapsed, "aracne")

TaxGroup (g, T_table = david_stool_lineages, g_tax = "Order", p_tax = "Genus")
}
