% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxGroup.R
\name{TaxGroup}
\alias{TaxGroup}
\title{Identifies the membership of a node to a certain higher-ranking taxon.}
\usage{
TaxGroup(g, T_table, g_tax, p_tax)
}
\arguments{
\item{g}{An igraph object.}

\item{T_table}{Taxa table as data.frame}

\item{g_tax}{The taxonomic level at which the nodes will be colored.}

\item{p_tax}{The taxonomic level at which the network was built.}
}
\value{
description
}
\description{
Identifies the membership of a node to a certain higher-ranking taxon.
}
\examples{
1+1

}
