% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxGroup.R
\name{TaxGroup}
\alias{TaxGroup}
\title{TaxGroup}
\usage{
TaxGroup(g, T_table, g_tax, p_tax)
}
\arguments{
\item{g}{An igraph object.}

\item{T_table}{Taxa table as data.frame}

\item{g_tax}{The taxonomic level at which the nodes will be colored.}

\item{p_tax}{The taxonomic level at which the network was built.}
}
\value{
Returns an igraph object with nodes colored based on their taxonomic rank.
}
\description{
This function identifies the membership of a node to a certain higher-ranking taxon.
}
\examples{

# The input can be a net pre-processed by net_inference
# This example uses david_stool data set

library (seqtime)
data ("tc_dstoolA")
g <- net_inference (tc_dstoolA, "aracne")

TaxGroup (g, T_table = david_stool_lineages, g_tax = "Order", p_tax = "Genus")
}
