% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_colored.R
\name{v_colored}
\alias{v_colored}
\title{Assign a color to each node of a network depending of a certain taxonomic level,
for example, the phylum.}
\usage{
v_colored(g, T_table, g_tax, p_tax, g_colors)
}
\arguments{
\item{g}{An igraph object.}

\item{T_table}{Taxa table as data.frame.}

\item{g_tax}{The taxonomic level at which the nodes will be colored.}

\item{p_tax}{The taxonomic level at which the network was built.}

\item{g_colors}{Set of colors. The vector must have the same length of g_tax elements.}
}
\value{
An igraph object.
}
\description{
Assign a color to each node of a network depending of a certain taxonomic level,
for example, the phylum.
}
\examples{

# Load this library to create a vector with colors
library(viridis)
c <- sample(viridis(152), 152)

data ("net_dstoolA")

g <- v_colored(net_dstoolA, T_table = david_stool_lineages, g_tax = "Class",
p_tax = "Family", g_colors = c)

plot(g)

}
